Add new option isTSX:
If this option isn't specified, the value is inferred from the file path.

If the file path also isn't specified, it is assumed that no TSX is being used.

File paths must end with '.tsx' if isTSX is specified to be 'true'.

Conversely setting isTSX to 'false' means file paths must end with '.mts'.

This is to make sure that the file path can later be used to determine if TSX is being used or not.